Create README.md
Meet Aditya Raj, a skilled frontend developer specializing in React.js. With a passion for creating engaging user experiences and a knack for crafting clean, efficient code, Aditya is at the forefront of web development innovation. Currently, he is  on an exciting journey, eagerly exploring new opportunities to further his expertise and make impactful contributions to the ever-evolving world of technology.